splite test call and check
