+++ src/plugins/azgenerator/AzExtensionIncrementalGenerator.ts
@@ -24,9 +24,9 @@ import { GenerateAzureCliParams } from "./templates/generated/CliParams";
 import { GenerateAzureCliValidators } from "./templates/generated/CliValidators";
 import {CliTestInit} from "./templates/tests/CliTestInit";
 import { CliTestPrepare } from "./templates/tests/CliTestPrepare";
-import { CliTestScenario, NeedPreparer } from "./templates/tests/CliTestScenario";
-import { inplaceGen } from "../../utils/inplace"
+import { CliTestScenario } from "./templates/tests/CliTestScenario";
 import { deepCopy } from '../../utils/helper';
+import { CliTestStep, NeedPreparer } from "./templates/tests/CliTestStep";
 
 export class AzExtensionIncrementalGenerator extends AzGeneratorBase {
     constructor(model: CodeModelAz, isDebugMode: boolean) {
@@ -81,6 +81,7 @@ export class AzExtensionIncrementalGenerator extends AzGeneratorBase {
         await this.generateIncrementalSingleAndAddtoOutput(new CliExtSetupPy(this.model, this.isDebugMode));
 
         await this.generateIncrementalSingleAndAddtoOutput(new CliTestInit(this.model, this.isDebugMode));
+        await this.generateFullSingleAndAddtoOutput(new CliTestStep(this.model, this.isDebugMode), true, true);
         let config: any = deepCopy(this.model.Extension_TestScenario);
         if(this.model.ConfiguredScenario){
             for (var ci = 0; ci < config.length; ci++) {
