+++ src/plugins/azgenerator/CodeModelAzImpl.ts
@@ -38,6 +38,7 @@ export class CodeModelCliImpl implements CodeModelAz {
     currentSubOptionIndex: number;
     paramActionNameReference: Map<Schema, string>;
     private _testScenario: any[];
+    private _defaultTestScenario: any[];
     private _configuredScenario: boolean;
     private _clientSubscriptionBound: boolean;
     private _clientBaseUrlBound: boolean;
@@ -508,6 +509,10 @@ export class CodeModelCliImpl implements CodeModelAz {
 
     public get Extension_TestScenario(): any {
         return this._testScenario;
+    }
+
+    public get Extension_DefaultTestScenario(): any {
+        return this._defaultTestScenario;
     }
 
     public get Extension_ClientSubscriptionBound(): boolean {
@@ -2337,11 +2342,15 @@ export class CodeModelCliImpl implements CodeModelAz {
             this.resource_pool.setResourceDepends(this.CommandGroup_Key, depend_resources, depend_parameters, createdObjectNames);
         });
 
-        if (!this._configuredScenario && isNullOrUndefined(this._testScenario)) {
-            this._testScenario = GenerateDefaultTestScenario(this.GetAllExamples());
-            this._testScenario = GenerateDefaultTestScenarioByDependency(this.GetAllExamples(), this.resource_pool, this._testScenario);
+        if (isNullOrUndefined(this._defaultTestScenario)) {
+            this._defaultTestScenario = GenerateDefaultTestScenario(this.GetAllExamples());
+            this._defaultTestScenario = GenerateDefaultTestScenarioByDependency(this.GetAllExamples(), this.resource_pool, this._defaultTestScenario);
             this.SortExamplesByDependency();
-            PrintTestScenario(this._testScenario);
+            PrintTestScenario(this._defaultTestScenario);
+        }
+
+        if (!this._configuredScenario && isNullOrUndefined(this._testScenario)) {
+            this._testScenario = this._defaultTestScenario;
         }
 
         let commandParams = {};
@@ -2421,26 +2430,25 @@ export class CodeModelCliImpl implements CodeModelAz {
 
         let scenarioExamples: Map<string, CommandExample> = new Map<string, CommandExample>();
         let commandExamples = this.GetAllExamples();
-        for (let i = 0; i < this._testScenario.length; i++) {
+        for (let i = 0; i < this._defaultTestScenario.length; i++) {
             for (let commandExample of commandExamples) {
-                if (this.matchExample(commandExample, this._testScenario[i]['name'])) {
-                    scenarioExamples.set(this._testScenario[i]['name'], commandExample);
+                if (this.matchExample(commandExample, this._defaultTestScenario[i]['name'])) {
+                    scenarioExamples.set(this._defaultTestScenario[i]['name'], commandExample);
                     break;
                 }
             }
-
         }
 
         let i = 0;
         let swapped = new Set<string>();    //for loop detecting
-        while (i < this._testScenario.length) {
-            for (let j = i + 1; j < this._testScenario.length; j++) {
+        while (i < this._defaultTestScenario.length) {
+            for (let j = i + 1; j < this._defaultTestScenario.length; j++) {
                 let swapId = `${i}<->${j}`;
                 if (swapped.has(swapId)) continue; // has loop, ignore the compare.
-                if (compare(scenarioExamples.get(this._testScenario[i]['name']), scenarioExamples.get(this._testScenario[j]['name'])) > 0) {
-                    let tmp = this._testScenario[i];
-                    this._testScenario[i] = this._testScenario[j];
-                    this._testScenario[j] = tmp;
+                if (compare(scenarioExamples.get(this._defaultTestScenario[i]['name']), scenarioExamples.get(this._defaultTestScenario[j]['name'])) > 0) {
+                    let tmp = this._defaultTestScenario[i];
+                    this._defaultTestScenario[i] = this._defaultTestScenario[j];
+                    this._defaultTestScenario[j] = tmp;
                     swapped.add(swapId);
                     i--;
                     break;
@@ -2448,7 +2456,6 @@ export class CodeModelCliImpl implements CodeModelAz {
             }
             i++;
         }
-        //this._testScenario = MergeSort(this._testScenario,compare);
     }
 
     public GetAllMethods(command_group?: string, callback?: () => void): any[] {
