+++ src/plugins/azgenerator/templates/tests/CliTestScenario.ts
@@ -5,18 +5,12 @@
 import * as path from 'path';
 import { CodeModelAz, CommandExample } from "../../CodeModelAz"
 import { PreparerEntity, getResourceKey } from "./ScenarioTool"
+import { CliTestStep } from "./CliTestStep"
 import { ToMultiLine, deepCopy } from '../../../../utils/helper';
 import { HeaderGenerator } from "../../Header";
 import { TemplateBase } from "../TemplateBase";
 import { PathConstants } from "../../../models";
 
-let usePreparers = false;
-let nameMap = {};
-let nameSeq = {};
-
-export function NeedPreparer(): boolean {
-    return usePreparers;
-}
 
 export class CliTestScenario extends TemplateBase {
     constructor(model: CodeModelAz, isDebugMode: boolean, testFilename: string, configValue:any) {
@@ -57,7 +51,6 @@ export class CliTestScenario extends TemplateBase {
 
         header.addImport("os");
         head.push("from azure.cli.testsdk import ScenarioTest");
-        head.push("from .. import try_manual, raise_if, calc_coverage");
         steps.push("");
         steps.push("");
         steps.push("TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))");
@@ -79,36 +72,20 @@ export class CliTestScenario extends TemplateBase {
 
         let imports: string[] = [];
         let decorators: string[] = [];
-        let parameterNames = this.InitiateDependencies(model, imports, decorators, initiates);
+        let parameterNames = CliTestStep.InitiateDependencies(model, imports, decorators, initiates);
         let jsonAdded = false;
-
-        function parameterLine(withChecksDef=false) {
-            let ret = "";
-            var paramList: string[] = deepCopy(parameterNames) as string[];
-            if (withChecksDef) {
-                paramList.push("checks=None");
-            }
-            for (let name of [...parameterNames, "checks=None"]) {
-                ret += `, ${name}`;
-            }
-            return ret;
-        }
 
         funcScenario.push("# Testcase");
         funcScenario.push("@try_manual");
-        funcScenario.push(...ToMultiLine(`def call_scenario(test${parameterLine()}):`));
+        funcScenario.push(...ToMultiLine(`def call_scenario(test${CliTestStep.parameterLine(parameterNames)}):`));
         model.GetResourcePool().clearExampleParams();
 
         // go through the examples to generate steps
         for (var ci = 0; ci < config.length; ci++) {
             let exampleId: string = config[ci].name;
-            let functionName: string = this.ToFunctionName(config[ci]);
+            let functionName: string = CliTestStep.ToFunctionName(config[ci]);
             if (exampleId) {
                 let disabled: string = config[ci].disabled ? "# " : "";
-                steps.push("# EXAMPLE: " + exampleId);
-                steps.push("@try_manual");
-                steps.push(...ToMultiLine(`def ${functionName}(test${parameterLine(true)}):`));
-
                 // find example by name
                 let found = false;
                 let examples: CommandExample[] = [];
@@ -116,194 +93,52 @@ export class CliTestScenario extends TemplateBase {
                 for (let exampleCmd of model.FindExampleById(exampleId, commandParams, examples)) {
                     if (exampleCmd && exampleCmd.length > 0) {
                         found = true;
-                        if (exampleCmd[0].indexOf(' delete') > -1) {
-                            exampleCmd[0] += " -y";
-                        }
-
                         let checks = model.GetExampleChecks(examples[exampleIdx++]);
                         if (checks.length > 0) {
-                            steps.push("    " + disabled + "_checks=[");
+                            funcScenario.push(...ToMultiLine(`    ${disabled}${functionName}(test${CliTestStep.parameterLine(parameterNames)}, checks=[`));
                             for (let check of checks) {
-                                ToMultiLine("    " + disabled + "    " + check, steps);
+                                ToMultiLine("    " + disabled + "    " + check, funcScenario);
                                 if (!jsonAdded && !disabled && check.indexOf("json.loads") >= 0) {
                                     header.addImport("json");
                                     jsonAdded = true;
                                 }
                             }
-                            steps.push("    " + disabled + "]");
+                            funcScenario.push(`    ${disabled}])`);
                         }
                         else {
-                            steps.push("    " + disabled + "_checks=[]");
+                            funcScenario.push(...ToMultiLine(`    ${functionName}(test${CliTestStep.parameterLine(parameterNames)}, checks=[])`));
                         }
-                        steps.push(`    if checks is not None:`);
-                        steps.push(`        _checks = checks`);
-
-                        for (let idx = 0; idx < exampleCmd.length; idx++) {
-                            let prefix: string = "    " + disabled + ((idx == 0) ? "test.cmd('" : "         '");
-                            let postfix: string = (idx < exampleCmd.length - 1) ? " '" : "',";
-                            ToMultiLine(prefix + exampleCmd[idx] + postfix, steps);
-                        }
-                        steps.push("    " + disabled + "         checks=_checks)");
-                        // let checks = model.GetExampleChecks(examples[exampleIdx++]);
-                        // if (checks.length > 0) {
-                        //     steps.push("    " + disabled + "         checks=[");
-                        //     for (let check of checks) {
-                        //         ToMultiLine("    " + disabled + "             " + check, steps);
-                        //         if (!jsonAdded && !disabled && check.indexOf("json.loads") >= 0) {
-                        //             header.addImport("json");
-                        //             jsonAdded = true;
-                        //         }
-                        //     }
-                        //     steps.push("    " + disabled + "         ])");
-                        // }
-                        // else {
-                        //     steps.push("    " + disabled + "         checks=[])");
-                        // }
                     }
                 }
-                if (!found) {
-                    steps.push("    # EXAMPLE NOT FOUND!");
-                    steps.push("    pass");
+                if (found) {
+                    head.push(`from example_steps import ${functionName}`);
                 }
                 else {
-                    for (let exampleCmd of model.FindExampleWaitById(exampleId)) {
-                        if (exampleCmd && exampleCmd.length > 0) {
-                            found = true;
-                            for (let idx = 0; idx < exampleCmd.length; idx++) {
-                                let prefix: string = "    " + disabled + ((idx == 0) ? "test.cmd('" : "         '");
-                                let postfix: string = (idx < exampleCmd.length - 1) ? " '" : "',";
-                                ToMultiLine(prefix + exampleCmd[idx] + postfix, steps);
-                            }
-                            steps.push("    " + disabled + "         checks=[])");
-                        }
-                    }
+                    funcScenario.push(...ToMultiLine(`    # STEP NOT FOUND: ${exampleId}`));
                 }
-                if (disabled) {
-                    steps.push("    pass");
-                }
-                steps.push("");
-                steps.push("");
-                funcScenario.push(...ToMultiLine(`    ${functionName}(test${parameterLine()})`));
             }
-            else if (functionName) {
+            else {
                 steps.push(`# Env ${functionName}`);
                 steps.push("@try_manual");
-                steps.push(...ToMultiLine(`def ${functionName}(test${parameterLine()}):`));
+                steps.push(...ToMultiLine(`def ${functionName}(test${CliTestStep.parameterLine(parameterNames)}):`));
                 steps.push("    pass");
                 steps.push("");
                 steps.push("");
-                funcScenario.push(...ToMultiLine(`    ${functionName}(test${parameterLine()})`));
-            }
+                funcScenario.push(...ToMultiLine(`    ${functionName}(test${CliTestStep.parameterLine(parameterNames)})`));
+            }   
         }
         funcScenario.push("");
         funcScenario.push("");
-        body.push(`        call_scenario(self${parameterLine()})`);
+        body.push(`        call_scenario(self${CliTestStep.parameterLine(parameterNames)})`);
         body.push(`        calc_coverage(__file__)`);
         body.push(`        raise_if()`);
         body.push("");
 
+        head.push("from .. import try_manual, raise_if, calc_coverage");
         let output = head.concat(imports, steps, funcScenario, class_info, decorators, initiates, body);
         output.forEach(element => {
             if (element.length > 120) header.disableLineTooLong = true;
         });
         return header.getLines().concat(output);
-    }
-
-    private InitiateDependencies(model: CodeModelAz, imports: string[], decorators: string[], initiates: string[]): string[] {
-        let decorated = [];
-        let internalObjects = [];
-        let parameterNames = [];
-        for (let entity of (model.GetPreparerEntities() as PreparerEntity[])) {
-            if (!entity.info.name) {
-                internalObjects.push([entity.info.class_name, getResourceKey(entity.info.class_name, entity.object_name), entity.info.createdObjectNames.indexOf(entity.object_name) >= 0, entity.object_name]);
-                continue;
-            }
-
-            // create preparers for outside dependency
-            let line: string = `    @${entity.info.name}(name_prefix='clitest${model.Extension_NameUnderscored}_${entity.object_name}'[:7], key='${getResourceKey(entity.info.class_name, entity.object_name)}'`;
-            for (let i = 0; i < entity.depend_parameter_values.length; i++) {
-                line += `, ${entity.info.depend_parameters[i]}='${entity.depend_parameter_values[i]}'`
-            }
-            if (entity.info.name == 'ResourceGroupPreparer') {
-                let parameterName = getResourceKey(entity.info.class_name, entity.object_name);
-                line += `, parameter_name='${parameterName}'`;
-                parameterNames.push(parameterName);
-            }
-            line += ")";
-            ToMultiLine(line, decorators);
-            if (decorated.indexOf(entity.info.name) < 0) {
-                if (entity.info.name == 'ResourceGroupPreparer') {
-                    imports.push(`from azure.cli.testsdk import ${entity.info.name}`);
-                }
-                else if (entity.info.name == 'StorageAccountPreparer') {
-                    imports.push(`from azure.cli.testsdk import ${entity.info.name}`);
-                } else {
-                    imports.push(`from .preparers import ${entity.info.name}`);
-                    usePreparers = true;
-                }
-                decorated.push(entity.info.name);
-            }
-        }
-        let funcLine = "    def test_" + model.Extension_NameUnderscored + "(self";
-        for (let parameterName of parameterNames) {
-            funcLine += `, ${parameterName}`;
-        }
-        funcLine += "):";
-        initiates.unshift(...ToMultiLine(funcLine));
-
-        // randomize name for internal resources
-        if (internalObjects.length > 0) {
-            initiates.push("        self.kwargs.update({");
-            for (let [class_name, kargs_key, hasCreateExample, object_name] of internalObjects) {
-                if (hasCreateExample && model.RandomizeNames) {
-                    const RANDOMIZE_MIN_LEN = 4;
-                    let prefixLen = Math.floor(object_name.length / 2);
-                    if (object_name.length - prefixLen < RANDOMIZE_MIN_LEN) prefixLen = Math.max(object_name.length - RANDOMIZE_MIN_LEN, 0);
-                    ToMultiLine(`            '${kargs_key}': self.create_random_name(prefix='${object_name}'[:${prefixLen}], length=${Math.max(object_name.length, RANDOMIZE_MIN_LEN)}),`, initiates);
-                }
-                else
-                    initiates.push(`            '${kargs_key}': '${object_name}',`);   // keep the original name in example if there is no create example in the test-scenario
-            }
-            initiates.push("        })");
-            initiates.push("");
-        }
-        return parameterNames;
-    }
-
-    private ToFunctionName(step: any): string {
-        let ret = undefined;
-        if (step.name)
-            ret = "step_" + step.name;
-        else if (step.function)
-            ret = step.function;
-        if (!ret) return undefined;
-
-        let funcname = "";
-        var letterNumber = /^[0-9a-zA-Z]+$/;
-        ret = (ret as string).toLowerCase();
-        for (let i = 0; i < ret.length; i++) {
-            funcname += ret[i].match(letterNumber) ? ret[i] : '_';
-        }
-        if (funcname.length > 50) {
-            if (!nameMap.hasOwnProperty(funcname)) {
-                let arr = funcname.split("_");
-                let shortName = arr.join("_");
-                if (arr.length > 4) {
-                    shortName = arr.slice(0, 4).join("_");
-                }
-                if (shortName.length > 50) shortName = shortName.substr(0, 50);
-
-                if (nameSeq.hasOwnProperty(shortName)) {
-                    nameSeq[shortName] += 1;
-                    nameMap[funcname] = shortName + nameSeq[shortName];
-                }
-                else {
-                    nameSeq[shortName] = 1;
-                    nameMap[funcname] = shortName;
-                }
-            }
-            funcname = nameMap[funcname];
-        }
-        return funcname;
     }
 }
